+++ cmd/substreams/run.go
@@ -28,6 +28,7 @@ func init() {
 	runCmd.Flags().Bool("final-blocks-only", false, "Only process blocks that have pass finality, to prevent any reorg and undo signal by staying further away from the chain HEAD")
 	runCmd.Flags().Bool("insecure", false, "Skip certificate validation on GRPC connection")
 	runCmd.Flags().Bool("plaintext", false, "Establish GRPC connection in plaintext")
+	runCmd.Flags().Bool("noop-mode", false, "Sends the request to the server with 'noop-mode', which will not send actual data, only populate the cache")
 	runCmd.Flags().StringP("output", "o", "", "Output mode, one of: [ui, json, jsonl, clock] Defaults to 'ui' when in a TTY is present, and 'json' otherwise")
 	runCmd.Flags().StringSlice("debug-modules-initial-snapshot", nil, "List of 'store' modules from which to print the initial data snapshot (Unavailable in Production Mode)")
 	runCmd.Flags().StringSlice("debug-modules-output", nil, "List of modules from which to print outputs, deltas and logs (Unavailable in Production Mode)")
@@ -173,6 +174,10 @@ func runRun(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("stop block: %w", err)
 	}
 
+	noopMode := sflags.MustGetBool(cmd, "noop-mode")
+	if noopMode && !productionMode {
+		zlog.Warn("noop-mode used without production-mode: server will execute in development mode without sending the data, this is probably not what you want")
+	}
 	req := &pbsubstreamsrpc.Request{
 		StartBlockNum:                       startBlock,
 		StartCursor:                         cursorStr,
@@ -182,6 +187,7 @@ func runRun(cmd *cobra.Command, args []string) error {
 		OutputModule:                        outputModule,
 		ProductionMode:                      productionMode,
 		DebugInitialStoreSnapshotForModules: debugModulesInitialSnapshot,
+		NoopMode:                            noopMode,
 	}
 
 	if err := req.Validate(); err != nil {